+++ packages/fluentui/react-northstar/src/components/Dropdown/Dropdown.tsx
@@ -122,6 +122,9 @@ export interface DropdownProps extends UIComponentProps<DropdownProps, DropdownS
   /** A dropdown can be formatted to appear inline next to other elements. */
   inline?: boolean;
 
+  /** A dropdown can have inverted colors. */
+  inverted?: boolean;
+
   /** Array of props for generating list options (Dropdown.Item[]) and selected item labels (Dropdown.SelectedItem[]), if it's a multiple selection. */
   items?: ShorthandCollection<DropdownItemProps>;
 
@@ -247,6 +250,7 @@ export interface DropdownState {
   value: ShorthandCollection<DropdownItemProps>;
   itemIsFromKeyboard: boolean;
   isFromKeyboard: boolean;
+  inverted: boolean;
 }
 
 class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, DropdownState> {
@@ -289,6 +293,7 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
     highlightFirstItemOnOpen: PropTypes.bool,
     highlightedIndex: PropTypes.number,
     inline: PropTypes.bool,
+    inverted: PropTypes.bool,
     items: customPropTypes.collectionShorthand,
     itemToString: PropTypes.func,
     itemToValue: PropTypes.func,
@@ -369,6 +374,7 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
       value: [],
       itemIsFromKeyboard: false,
       isFromKeyboard: false,
+      inverted: false,
     };
   }